requirements.txt
folium==0.14.0
haversine==2.7.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
plotly==5.11.0
streamlit==1.16.0
streamlit-folium==0.9.0